Move column hook to shared
