+++ packages/react/src/verbs/Recode/hooks.ts
@@ -2,13 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import {
-	RecodeStep,
-	RecodeArgs,
-	Value,
-	DataType,
-	columnType,
-} from '@data-wrangling-components/core'
+import { RecodeStep, RecodeArgs, Value } from '@data-wrangling-components/core'
 import { op } from 'arquero'
 import ColumnTable from 'arquero/dist/types/table/column-table'
 import { useCallback, useMemo } from 'react'
@@ -28,15 +22,6 @@ export function useColumnValues(
 		}
 		return table.orderby(column).rollup(args).objects()[0][column] as Value[]
 	}, [table, internal])
-}
-
-export function useColumnType(table?: ColumnTable, column?: string): DataType {
-	return useMemo(() => {
-		if (!table || !column) {
-			return DataType.Unknown
-		}
-		return columnType(table, column)
-	}, [table, column])
 }
 
 export function useHandleRecodeChange(
